refactor: `lib/constant`에서 api url를 분리한다 (#134)
